+++ react/utils/client/links/uriSwitchLink.ts
@@ -53,17 +53,29 @@ const extractHints = (query: ASTNode, meta: CacheHints) => {
   }
 }
 
-export const createUriSwitchLink = (baseURI: string, workspace: string, locale: string, domain?: string) =>
+export const createUriSwitchLink = (baseURI: string, workspace: string, locale: string, production: boolean, domain?: string) =>
   new ApolloLink((operation: Operation, forward?: NextLink) => {
     operation.setContext((oldContext: OperationContext) => {
       const { fetchOptions = {}, runtime: {appsEtag, cacheHints} } = oldContext
-      const oldMethod = fetchOptions.method || 'POST'
       const hash = generateHash(operation.query)
+
+      if (!production && !hash) {
+        throw new Error(
+          'Could not generate hash from query. Are you using graphql-tag ? Split your graphql queries in .graphql files and import them instead'
+        )
+      }
+
+      const includeQuery = (oldContext as any).http?.includeQuery || !hash
+      const oldMethod = includeQuery ? 'POST' : (fetchOptions.method || 'POST')
       const protocol = canUseDOM ? 'https:' : 'http:'
       const {maxAge, scope, version, operationType} = extractHints(operation.query, cacheHints[hash])
       const method = (equals(scope, 'private') && equals(operationType, 'query')) ? 'POST' : oldMethod
       return {
         ...oldContext,
+        http: {
+          ...(oldContext as any).http,
+          includeQuery,
+        },
         fetchOptions: {...fetchOptions, method},
         uri: `${protocol}//${baseURI}/_v/${scope}/graphql/v${version}?workspace=${workspace}&maxAge=${maxAge}&appsEtag=${appsEtag}&domain=${domain}&locale=${locale}`,
       }