+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Expose `useExperimentalLazyImages`
+- Expose `useExperimentalLazyImagesContext`
 
 ## [8.98.2] - 2020-04-06
 ### Fixed