+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- PersistedQueryNotFoundError
+  - Adds query to request when persisted query is not found to avoid PersistedQueryNotFoundError from graphql-server
+  - Uses POST on those requests
+  - Throws error in development when persisted query is not found
 
 ## [8.99.1] - 2020-04-24
 ### Fixed